get sample response successfully

+++ src/main/kotlin/com/example/ordrsmartapi/repository/IProductRepository.kt
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface IProductRepository : CrudRepository<Product, Int> {
+interface IProductRepository : CrudRepository<Product, Long> {
 }
